+++ src/Ydb.Sdk/test/Ydb.Sdk.Ado.Tests/YdbCommandTests.cs
@@ -373,6 +373,10 @@ public class YdbCommandTests : TestBase
                             new[] { -maxInterval64, maxInterval64 })
                     }
                 }.ExecuteScalarAsync());
+
+                await Task.Delay(100_000);
+                
+                
             }
         );